Refactored PassThroughOptionParser to parser.py
